fix deprecated usages of Clickable Text and SwipeState.progress APIs, turn swipeProgress param to lambda in HomeScreen; fix Ios platform version logic

+++ composeApp/src/commonMain/kotlin/platform/Platform.kt
@@ -2,7 +2,7 @@ package platform
 
 sealed interface Platform {
     data class Android(val version: Int) : Platform
-    data class Ios(val version: Int) : Platform
+    data class Ios(val version: Double) : Platform
     data object Desktop : Platform
 
     sealed interface Web : Platform {
